requirements.txt
asgiref==3.2.10
dj-database-url==0.5.0
Django==3.1
django-heroku==0.3.1
gunicorn==20.0.4
psycopg2==2.8.5
pytz==2020.1
sqlparse==0.3.1
whitenoise==5.2.0
pylint-django==2.3.0
django-registration==3.1
six==1.15.0
validate-email==1.3
django-crispy-forms==1.9.2
Pillow==7.2.0
django-ckeditor==6.0.0
django-weekday-field==1.1.0
django-filter==2.3.0
django-extensions==3.0.8